+++ src/uncertain_statistics/statsbase/core_stats_uncertaindatasets.jl
@@ -64,17 +64,39 @@ Statistics.var(d::AbstractUncertainValueDataset, n::Int; kwargs...) =
 """
     cor(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int; kwargs...)
 
-Compute the Pearson correlation between two `AbstractUncertainValueDataset`s by realising
-both datasets `n` times.
+Obtain a distribution for the Pearson correlation between two uncertain datasets `d1` and `d2`.
+This is done by resampling both datasets multiple times and compute the correlation between 
+those draws. This yields a distribution of correlation estimates. 
+
+The procedure is as follows. 
+
+1. First, draw a realisation of `d1` according to the distributions furnishing its uncertain values. 
+2. Then, draw a realisation `d2` according to its furnishing distributions. 
+3. Compute the correlation between those two draws/realisations, both of which are vectors of 
+    length `L = length(d1) = length(d2)`. 
+4. Repeat the procedure `n` times, drawing `n` separate pairs of realisations of `d1` and `d2`.
+    This yields `n` estimates of the correlation between `d1` and `d2`, which is returned as 
+    a vector.
 """
-Statistics.cor(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int; kwargs...) =
+Statistics.cor(d1::DT, d2::DT, n::Int; kwargs...) where {DT <: AbstractUncertainValueDataset} =
     [cor(resample(d1), resample(d2), kwargs...) for i = 1:n]
 
 """
     cov(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int; kwargs...)
 
-Compute the covariance between two `AbstractUncertainValueDataset`s by realising
-both datasets `n` times.
+Obtain a distribution for the covariance between two uncertain datasets `d1` and `d2`.
+This is done by resampling both datasets multiple times and compute the covariance between 
+those draws. This yields a distribution of covariance estimates. 
+
+The procedure is as follows. 
+
+1. First, draw a realisation of `d1` according to the distributions furnishing its uncertain values. 
+2. Then, draw a realisation `d2` according to its furnishing distributions. 
+3. Compute the covariance between those two draws/realisations, both of which are vectors of 
+    length `L = length(d1) = length(d2)`. 
+4. Repeat the procedure `n` times, drawing `n` separate pairs of realisations of `d1` and `d2`.
+    This yields `n` estimates of the covariance between `d1` and `d2`, which is returned as 
+    a vector.
 """
 Statistics.cov(d1::AbstractUncertainValueDataset, d2::AbstractUncertainValueDataset, n::Int; kwargs...) =
     [cov(resample(d1), resample(d2), kwargs...) for i = 1:n]